add favorite

+++ routes/user_api.py
@@ -2,10 +2,12 @@ from flask import Blueprint, jsonify, request
 from models.user import User, UserSchema
 from models.favorite import Favorite
 from extension import db
+from models.account import Account
 from genetic_algorithm.GA import genetic_algorithm, print_menu
 from flask_jwt_extended import jwt_required, create_access_token, get_jwt_identity
 from genetic_algorithm.genetic_algorithm import GeneticAlgorithm
 from models.dish import Dish, DishSchema
+from ultralytics import YOLO
 
 user_api = Blueprint('user_api',__name__,url_prefix='/api/user')
 
@@ -14,23 +16,33 @@ users_schema = UserSchema(many=True)
 dish_schema = DishSchema()
 dishs_schema = DishSchema(many=True)
 
-@user_api.route("/get", methods=["GET"])
-def users_list():
-    users = User.query.all()
-    return users_schema.dump(users)
+# @user_api.route("/get", methods=["GET"])
+# def users_list():
+#     users = User.query.all()
+#     return users_schema.dump(users)
 
-@user_api.route("/get/<id>", methods=["GET"])
-def user_detail(id):
-    user = User.query.get(id)
+@user_api.route("/get", methods=["GET"])
+@jwt_required()
+def user_detail():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
-        return jsonify({'error': 'User not found'}), 404
+        return jsonify({'message': 'User not found'}), 404
     user_details = user.get_user_details()
     return jsonify(user_details)
 
 
-@user_api.route("/update_username/<id>", methods=["PUT"])
-def username_update(id):
-    user = User.query.get(id)
+@user_api.route("/update_username", methods=["PUT"])
+@jwt_required()
+def username_update():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
         return jsonify({'error': 'User not found'}), 404
     new_username=request.json["username"]
@@ -38,9 +50,14 @@ def username_update(id):
     db.session.commit()
     return user_schema.jsonify(user)
 
-@user_api.route("/update_age_height_weight/<id>", methods=["PUT"])
-def age_height_weight_update(id):
-    user = User.query.get(id)
+@user_api.route("/update_age_height_weight", methods=["PUT"])
+@jwt_required()
+def age_height_weight_update():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
         return jsonify({'error': 'User not found'}), 404
     new_age=request.json["age"]
@@ -52,9 +69,14 @@ def age_height_weight_update(id):
     db.session.commit()
     return user_schema.jsonify(user)
 
-@user_api.route("/update_aim/<id>", methods=["PUT"])
-def aim_update(id):
-    user = User.query.get(id)
+@user_api.route("/update_aim", methods=["PUT"])
+@jwt_required()
+def aim_update():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
         return jsonify({'error': 'User not found'}), 404
     new_aim=request.json["aim"]
@@ -62,9 +84,14 @@ def aim_update(id):
     db.session.commit()
     return user_schema.jsonify(user)
 
-@user_api.route("/update_gender/<id>", methods=["PUT"])
-def gender_update(id):
-    user = User.query.get(id)
+@user_api.route("/update_gender", methods=["PUT"])
+@jwt_required()
+def gender_update():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
         return jsonify({'error': 'User not found'}), 404
     new_gender=request.json["gender"]
@@ -72,53 +99,49 @@ def gender_update(id):
     db.session.commit()
     return user_schema.jsonify(user)
 
-@user_api.route("/update_exercise/<id>", methods=["PUT"])
-def exercise_update(id):
-    user = User.query.get(id)
+@user_api.route("/update_exercise", methods=["PUT"])
+@jwt_required()
+def exercise_update():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'msg': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
-        return jsonify({'error': 'User not found'}), 404
+        return jsonify({'msg': 'User not found'}), 404
     new_exercise=request.json["exercise"]
     user.exercise = new_exercise
     db.session.commit()
     return user_schema.jsonify(user)
 
-@user_api.route("/delete/<id>", methods=["DELETE"])
+@user_api.route("/delete", methods=["DELETE"])
 @jwt_required()
-def user_delete(id):
-    user = User.query.get(id)
+def user_delete():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
         return jsonify({'error': 'User not found'}), 404
     db.session.delete(user)
     db.session.commit()
     return user_schema.jsonify(user)
 
-@user_api.route('/add_favorite/<id>', methods=["POST"])
-def add_favorite_dish(id):
-    user = User.query.get(id)
-    if not user:
-        return jsonify({'message': 'User not found'}), 404
-    new_favorite_dish_id = request.json["dish_id"]
-    dish = Dish.query.get(new_favorite_dish_id)
-    if not dish:
-        return jsonify({"message": "Món ăn không tồn tại"}), 404
-    existing_favorite = (db.session.query(Favorite)
-                         .filter_by(dish_id=new_favorite_dish_id, user_id=id).first())
-    if existing_favorite:
-        return jsonify({'message': 'Món này đã ở trong danh sách yêu thích, bạn có muốn xóa?'}), 201
-    else:
-        new_favorite = Favorite(id, new_favorite_dish_id)
-        db.session.add(new_favorite)
-        db.session.commit()
-        return jsonify({'message': 'Thêm thành công'}), 200
-
 @user_api.route('/run_genetic_algorithm', methods=["GET"])
+@jwt_required()
 def run_genetic_algorithm():
     result = genetic_algorithm()
     return jsonify("")
 
-@user_api.route('/new_genetic_algorithm/<id>', methods=["GET"])
-def new_genetic_algorithm(id):
-    user = User.query.get(id)
+@user_api.route('/new_genetic_algorithm', methods=["GET"])
+@jwt_required()
+def new_genetic_algorithm():
+    current_user_email = get_jwt_identity()
+    user_account = Account.query.filter_by(email=current_user_email).first()
+    if not user_account:
+        return jsonify({'message': 'Unauthorized'}), 404
+    user = User.query.get(user_account.user_id)
     if not user:
         return jsonify({'message': 'Không tồn tại người dùng này'}), 404
     if not user.is_validate_user_data():
@@ -126,4 +149,11 @@ def new_genetic_algorithm(id):
     GA = GeneticAlgorithm(6, 0.2, 5, id)
     a = GA.main_genetic_algorithm()
     return a
+
+@user_api.route('/detect_ingredient', methods=['POST'])
+def detect_ingredient():
+    model = YOLO('yolov9c.pt')
+    model.info()
+    results = model.train(data='coco8.yaml', epochs=100, imgsz=640)
+    results = model('path/to/bus.jpg')
